Move apiversionMiddleware to headerMiddleware
This is more general/useful, and not that much more difficult to use in
the case of the API-Version header.
